refactor: split grace_hash_join.rs into submodules
Move the 2837-line operator into grace_hash_join/:
- mod.rs: GraceHashJoinExec, constants, ExecutionPlan impl, tests
- exec.rs: fast/slow path, per-partition joins, recursive repartitioning
- partition.rs: ScratchSpace, HashPartition, Phase 1/2 partitioners, spill
  helpers, merge logic
- spill.rs: SpillWriter, SpillReaderExec, StreamSourceExec
- metrics.rs: GraceHashJoinMetrics

Also fixes a latent clippy clone_on_ref_ptr violation on properties().clone().
